feat: Moving controller code in trait
